[SDKS-5498] Move event send uri to EventSender and add test cases
