+++ client/src/main/java/io/split/client/events/EventsSender.java
@@ -1,37 +1,44 @@
 package io.split.client.events;
 
+import com.google.common.annotations.VisibleForTesting;
 import io.split.client.dtos.Event;
+import io.split.client.utils.Utils;
 import io.split.service.HttpPostImp;
-import io.split.telemetry.domain.enums.HTTPLatenciesEnum;
 import io.split.telemetry.domain.enums.HttpParamsWrapper;
-import io.split.telemetry.domain.enums.LastSynchronizationRecordsEnum;
-import io.split.telemetry.domain.enums.ResourceEnum;
 import io.split.telemetry.storage.TelemetryRuntimeProducer;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
 
 import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.List;
 
 import static com.google.gson.internal.$Gson$Preconditions.checkNotNull;
 
 public class EventsSender {
-    private final URI _endpoint;
+
+    private static final String BULK_ENDPOINT_PATH = "api/events/bulk";
+    private final URI _bulkEndpoint;
     private final CloseableHttpClient _client;
     private final TelemetryRuntimeProducer _telemetryRuntimeProducer;
     private final HttpPostImp _httpPostImp;
 
-    public static EventsSender create(CloseableHttpClient httpclient, URI eventsTarget, TelemetryRuntimeProducer telemetryRuntimeProducer) {
-        return new EventsSender(httpclient, eventsTarget, telemetryRuntimeProducer);
+    public static EventsSender create(CloseableHttpClient httpclient, URI eventsTarget, TelemetryRuntimeProducer telemetryRuntimeProducer) throws URISyntaxException {
+        return new EventsSender(httpclient, Utils.appendPath(eventsTarget, BULK_ENDPOINT_PATH), telemetryRuntimeProducer);
     }
 
     EventsSender(CloseableHttpClient httpclient, URI eventsTarget, TelemetryRuntimeProducer telemetryRuntimeProducer) {
         _client = checkNotNull(httpclient);
-        _endpoint = checkNotNull(eventsTarget);
+        _bulkEndpoint = checkNotNull(eventsTarget);
         _telemetryRuntimeProducer = checkNotNull(telemetryRuntimeProducer);
         _httpPostImp = new HttpPostImp(httpclient, telemetryRuntimeProducer);
     }
 
     public void sendEvents(List<Event> _data) {
-        _httpPostImp.post(_endpoint, _data, "Events ", HttpParamsWrapper.EVENTS);
+        _httpPostImp.post(_bulkEndpoint, _data, "Events ", HttpParamsWrapper.EVENTS);
+    }
+
+    @VisibleForTesting
+    URI getBulkEndpoint() {
+        return _bulkEndpoint;
     }
 }
